requirements.txt
flask
groq
python-dotenv
sentence-transformers
torch